refactor: CATEGORY_LABELS定数を一元化し、Top3Page.tsxを分割する (#76, #69)
