+++ src/constants/category.ts
@@ -1,0 +1,19 @@
+import type { MediaCategory } from '../types/common'
+
+export const CATEGORY_LABELS_JA: Record<MediaCategory, string> = {
+  book: '書籍',
+  music: '音楽',
+  movie: '映画',
+}
+
+export const CATEGORY_LABELS_EN: Record<MediaCategory, string> = {
+  book: 'Book',
+  music: 'Music',
+  movie: 'Movie',
+}
+
+export const API_ENDPOINTS: Record<MediaCategory, string> = {
+  book: '/api/books',
+  music: '/api/music',
+  movie: '/api/movies',
+}
